remove GCR specific tests
there isn't enough of a benefit having a test that runs against one of
the three top IaaS's. I'm of the opinion that running the tests against
DockerHub is enough since they set and follow the image registry
standard. If it works on DockerHub I'd expect it to work everywhere
else, and if it doesn't then I'd bet it's the non-DockerHub platform
that is behaving weirdly

Signed-off-by: Taylor Silva <dev@taydev.net>